# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GradeFlow API
servers:
  - url: /api
    description: Base API path
security:
  - cookieAuth: []
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "400":
          description: Invalid input or wrong current password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /auth/request-password-reset:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestPasswordResetRequest" }
      responses:
        "200":
          description: |
            Always returns `{ ok: true }` regardless of whether the email
            matches a registered account — both the response shape and the
            response timing are equalized to prevent account enumeration.
            The reset link is NEVER returned in this response: it can only
            be delivered through a trusted side channel (email integration).
            When no email integration is configured, admins should use the
            admin-issued temporary password flow at
            `POST /users/{userId}/reset-password` instead.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordResetRequestResult" }
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      parameters:
        - in: query
          name: courseId
          schema: { type: integer }
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Assignment" }
    post:
      operationId: createAssignment
      tags: [assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assignment" }
  /assignments/{assignmentId}/close:
    parameters:
      - $ref: "#/components/parameters/AssignmentId"
    post:
      operationId: closeAssignment
      tags: [assignments]
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assignment" }
  /assignments/{assignmentId}/reopen:
    parameters:
      - $ref: "#/components/parameters/AssignmentId"
    post:
      operationId: reopenAssignment
      tags: [assignments]
      responses:
        "200":
          description: Reopened
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assignment" }
  /assignments/{assignmentId}:
    parameters:
      - $ref: "#/components/parameters/AssignmentId"
    get:
      operationId: getAssignment
      tags: [assignments]
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssignmentDetail" }
    patch:
      operationId: updateAssignment
      tags: [assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assignment" }
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /assignments/{assignmentId}/files:
    parameters:
      - $ref: "#/components/parameters/AssignmentId"
    get:
      operationId: listAssignmentFiles
      tags: [assignments]
      responses:
        "200":
          description: Files
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AssignmentFile" }
    post:
      operationId: attachAssignmentFile
      tags: [assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentFileInput" }
      responses:
        "201":
          description: Attached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssignmentFile" }
  /assignments/{assignmentId}/files/{fileId}:
    parameters:
      - $ref: "#/components/parameters/AssignmentId"
      - in: path
        name: fileId
        required: true
        schema: { type: integer }
    patch:
      operationId: updateAssignmentFile
      tags: [assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentFilePatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssignmentFile" }
    delete:
      operationId: deleteAssignmentFile
      tags: [assignments]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
...
  /export/assignment/{assignmentId}:
    parameters:
      - $ref: "#/components/parameters/AssignmentId"
    get:
      operationId: exportAssignmentCsv
      tags: [importExport]
      responses:
        "200":
          description: CSV
          content:
            text/csv:
              schema: { type: string }
...
  /assignment-files/{fileId}/preview:
    get:
      tags: [assignments]
      operationId: previewAssignmentFile
      parameters:
        - in: path
          name: fileId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: PDF stream rendered inline
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404":
          description: File not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "415":
          description: Preview not supported for this file type — clients should fall back to download
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PreviewUnsupportedError" }
  /assignment-files/{fileId}/download:
    get:
      tags: [assignments]
      operationId: downloadAssignmentFile
      parameters:
        - in: path
          name: fileId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Raw file download (Content-Disposition attachment)
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404":
          description: File not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gradeflow_token
      description: |
        Session JWT cookie set by `POST /auth/login`. All endpoints require it
        unless the operation explicitly overrides `security: []`.
  parameters:
...
    AssignmentId:
      in: path
      name: assignmentId
      required: true
      schema: { type: integer }
...
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    AuthUser:
      type: object
      required: [id, email, name, role]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string }
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 8 }
    RequestPasswordResetRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetRequestResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 8 }
...
    Assignment:
      type: object
      required: [id, courseId, name, maxScore, weight, gradingScale, closed]
      properties:
        id: { type: integer }
        courseId: { type: integer }
        courseName: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        dueDate: { type: string, format: date-time, nullable: true }
        maxScore: { type: number }
        weight: { type: number }
        gradingScale: { type: string }
        closed: { type: boolean }
        submissionCount: { type: integer }
        gradedCount: { type: integer }
    AssignmentDetail:
      allOf:
        - $ref: "#/components/schemas/Assignment"
        - type: object
          properties:
            files:
              type: array
              items: { $ref: "#/components/schemas/AssignmentFile" }
    AssignmentInput:
      type: object
      required: [courseId, name, maxScore, weight, gradingScale]
      properties:
        courseId: { type: integer }
        name: { type: string, minLength: 1 }
        description: { type: string, nullable: true }
        dueDate: { type: string, format: date-time, nullable: true }
        maxScore: { type: number }
        weight: { type: number }
        gradingScale: { type: string }
    AssignmentFile:
      type: object
      required: [id, assignmentId, name, objectPath, contentType, size, fileType]
      properties:
        id: { type: integer }
        assignmentId: { type: integer }
        name: { type: string }
        objectPath: { type: string }
        contentType: { type: string }
        size: { type: integer }
        fileType:
          type: string
          enum: [instructions, grading_guide, reference, extra]
    AssignmentFileInput:
      type: object
      required: [name, objectPath, contentType, size]
      properties:
        name: { type: string, minLength: 1 }
        objectPath: { type: string, minLength: 1 }
        contentType: { type: string, minLength: 1 }
        size: { type: integer, minimum: 0 }
        fileType:
          type: string
          enum: [instructions, grading_guide, reference, extra]
    AssignmentFilePatch:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        fileType:
          type: string
          enum: [instructions, grading_guide, reference, extra]
...
    PreviewUnsupportedError:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code:
          type: string
          enum: [PREVIEW_REQUIRES_DOWNLOAD, PREVIEW_UNSUPPORTED]